+++ lib/presentation/widgets/certifications_config/desktop_certifications_config.dart
@@ -1,0 +1,19 @@
+
+
+
+import 'package:accredit/presentation/widgets/certifications_config/base_certification_config.dart';
+import 'package:flutter/material.dart';
+
+class DesktopCertificationConfig extends BaseCertificationConfig {
+  DesktopCertificationConfig({super.key, required super.documentId});
+
+  @override
+  State<DesktopCertificationConfig> createState() => _DesktopCertificationConfigState();
+}
+
+class _DesktopCertificationConfigState extends BaseCertificationConfigState<DesktopCertificationConfig> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}