+++ lib/presentation/widgets/page_filter/desktop_page_filter.dart
@@ -2,6 +2,9 @@ import 'dart:typed_data';
 
 import 'package:accredit/core/utils/my_dialogs.dart';
 import 'package:accredit/core/utils/my_logs.dart';
+import 'package:accredit/core/utils/my_nagivation.dart';
+import 'package:accredit/domain/models/topic_identifications.dart';
+import 'package:accredit/presentation/widgets/certifications_config/on_certifications_config.dart';
 import 'package:accredit/presentation/widgets/page_filter/base_page_filter.dart';
 import 'package:flutter/material.dart';
 
@@ -40,7 +43,7 @@ class DesktopPageFilter extends StatelessWidget {
                 
           } else {
             // NavigationService.push(MultiPagePreviewScreen(selectedPages: selected));
-            debug('Selected pages: $selected');
+           NavigationService.push(OnCertificationConfigScreen(documentId:  documentId));
           }
         }
       },