+++ lib/presentation/widgets/page_filter/mobile_page_filter.dart
@@ -2,6 +2,9 @@ import 'dart:typed_data';
 
 import 'package:accredit/core/utils/my_dialogs.dart';
 import 'package:accredit/core/utils/my_logs.dart';
+import 'package:accredit/core/utils/my_nagivation.dart';
+import 'package:accredit/domain/models/topic_identifications.dart';
+import 'package:accredit/presentation/widgets/certifications_config/on_certifications_config.dart';
 import 'package:accredit/presentation/widgets/page_filter/base_page_filter.dart';
 import 'package:flutter/material.dart';
 
@@ -38,8 +41,7 @@ class MobilePageFilter extends StatelessWidget {
                 message: 'You can select up to 30 pages only. You selected ${selected.length} pages.');
                 
           } else {
-            // NavigationService.push(MultiPagePreviewScreen(selectedPages: selected));
-            debug('Selected pages: $selected');
+            NavigationService.push(OnCertificationConfigScreen(documentId:  documentId));
           }
         }
       },